chore(migration): move headstate into state/headstate
