+++ node/migration.go
@@ -11,7 +11,7 @@ import (
 	"github.com/NethermindEth/juno/migration"
 	"github.com/NethermindEth/juno/migration/blocktransactions"
 	"github.com/NethermindEth/juno/migration/deprecated" //nolint:staticcheck,nolintlint,lll // ignore statick check package will be removed in future, nolinlint because main config does not check
-	"github.com/NethermindEth/juno/migration/headstate"
+	"github.com/NethermindEth/juno/migration/state/headstate"
 	"github.com/NethermindEth/juno/migration/historyprunner"
 	"github.com/NethermindEth/juno/utils/log"
 )
